Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 int main() {
-    float a;
+    int a;
    
-    scanf("%f",&a);
-    printf("You entered: %f",a);
+    scanf("%d",&a);
+    printf("You entered: %d",a);
    
     return 0;
 }
